feat(stepper): allow fractional gear_ratio + enforce max_step_vel cap

+++ include/cmeresearch_stepper_driver/tinkerforge_bricklet_stepper_driver.hpp
@@ -43,7 +43,7 @@ class StepperDriver {
           int acceleration,
           int deceleration,
           int steps_per_revolution,
-          int gear_ratio,
+          double gear_ratio,
           int max_step_vel,
           int standstill_current = 200,
           int motor_run_current = 800,
@@ -86,7 +86,7 @@ class StepperDriver {
         // parameters set by launch file
         int step_resolution_;
         int steps_per_revolution_;
-        int gear_ratio_;
+        double gear_ratio_;
         int max_step_vel_;
         int acceleration_;
         int deceleration_;
